Инкремент 11. Проверка 007/022

+++ internal/postgresql/postgresql.go
@@ -99,7 +99,7 @@ func CreateTable(pool *pgx.Conn) {
 						"ID" character varying COLLATE pg_catalog."default",
 						"MType" character varying COLLATE pg_catalog."default",
 						"Value" double precision NOT NULL DEFAULT 0,
-						"Delta" integer NOT NULL DEFAULT 0,
+						"Delta" bigint NOT NULL DEFAULT 0,
 						"Hash" character varying COLLATE pg_catalog."default"
 					)
